some renaming and ordering
