+++ src/react/components/Popup.js
@@ -1,8 +1,9 @@
 import { Modal } from "react-bootstrap";
 import React from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
-import Button from "./Button";
 import "./Popup.css";
+
+import Button from "./Button";
 
 function Popup({
   show,
